moves parts of type_system to type_system_adjunct
