+++ bigraph_schema/__init__.py
@@ -3,6 +3,6 @@ from bigraph_schema.registry import (
     strip_schema_keys, type_parameter_key, non_schema_keys, set_path, transform_path)
 from bigraph_schema.utilities import get_path, visit_method
 from bigraph_schema.edge import Edge
-from bigraph_schema.type_system import TypeSystem, type_schema_keys, resolve_path
+from bigraph_schema.type_system import TypeSystem
 from bigraph_schema.protocols import local_lookup_module
-from bigraph_schema.type_functions import FUNCTION_TYPE, METHOD_TYPE
+from bigraph_schema.type_functions import FUNCTION_TYPE, METHOD_TYPE, type_schema_keys, resolve_path
